+++ app/page.tsx
@@ -1,7 +1,6 @@
-import Login from "@/app/Login/Login";
+import Login from "@/app/components/Login/Login";
 import classNames from "classnames";
 import Image from "next/image";
-import Link from "next/link";
 import styles from "./page.module.css";
 
 export default function Home() {
